Reformats; change mouse callback argument type.   remove minmax all; I guess windows c++ allows min/max

+++ include/render.h
@@ -325,7 +325,7 @@ typedef void (CPROC*HideAndRestoreCallback)( uintptr_t psvUser );
 typedef void (CPROC*RedrawCallback)( uintptr_t psvUser, PRENDERER self );
 /* function signature for the mouse callback  which can be specified to handle events from mouse motion on the display.  see SetMouseHandler.
   would be 'wise' to retun 0 if ignored, 1 if observed (perhaps not used), but NOT ignored.*/
-typedef int  (CPROC*MouseCallback)( uintptr_t psvUser, int32_t x, int32_t y, uint32_t b );
+typedef uintptr_t  (CPROC*MouseCallback)( uintptr_t psvUser, int32_t x, int32_t y, uint32_t b );
 
 typedef struct input_point
 {
